Created a copy of Index and a new style sheet
I made a copy of the index file so that I could modify it to make it look a bit more presentable and created a new stylesheet to go along with it.
This is not currently finished, but it had a color scheme I thought looked a bit nice and the startings of something.